requirements.txt
Flask==0.11.1
gunicorn==19.6.0
google-cloud==0.33.1